set DefaultMaxWakuMessageSizeStr as the only source of truth

+++ tests/waku_relay/test_protocol.nim
@@ -1007,7 +1007,7 @@ suite "Waku Relay":
       let
         msgWithoutPayload =
               fakeWakuMessage(contentTopic=contentTopic, payload=getByteSequence(0))
-        sizeEmptyMsg = msgWithoutPayload.encode().buffer.len
+        sizeEmptyMsg = uint64(msgWithoutPayload.encode().buffer.len)
 
       let
         msg1 = fakeWakuMessage(contentTopic=contentTopic, payload=getByteSequence(1024)) # 1KiB
